feat(gpt): GPT 일정 생성 서비스 연동

+++ trip_planner_back/src/main/java/com/korit/trip_planner_back/dto/tsp/TspRequestDto.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Builder
 public class TspRequestDto {
 
-    private List<Long> spotIds;
+    private List<Integer> spotIds;
     private double startLat;
     private double startLon;
     private Double endLat;
